% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{adjustIndexLength}
\alias{adjustIndexLength}
\title{adjustIndexLength (not used in paper)}
\usage{
adjustIndexLength(index, length = NULL, stratas = stratas)
}
\value{



}
\description{
adjustIndexLength (not used in paper)
}
