% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{constructIntPoints}
\alias{constructIntPoints}
\title{constructIntPoints}
\usage{
constructIntPoints(conf, confPred)
}
\arguments{
\item{conf}{Configurations for catch-at-length}

\item{confPred}{Prediction configurations}
}
\value{
A list with integration points. Strata number for each integration point is included.
}
\description{
constructIntPoints
}
