% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{createMesh}
\alias{createMesh}
\title{Create mesh}
\usage{
createMesh(conf)
}
\arguments{
\item{conf}{}
}
\value{
Mesh used
}
\description{
This function creates the mesh to represent the Matern covariance strucutre
}
