% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defConfPred}
\alias{defConfPred}
\title{defConfPred}
\usage{
defConfPred(conf, cellsize = 20, Depth = "Data")
}
\arguments{
\item{conf}{Configurations used when fitting the model}

\item{cellsize}{provide distance between integration points}

\item{Depth}{if "NOAA": use NOAA data base for estimating depth in integration points; if GEBCO (.nc) file: use file for estimating depth in integration}
}
\description{
Configurations used for prediction.
}
\details{
This function sets up the configurations for the predictions, i.e. the index-at-length and index-at-age.
}
