% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk_functions.R
\name{defpar_alk}
\alias{defpar_alk}
\title{Set up data}
\usage{
defpar_alk(data, conf)
}
\arguments{
\item{data}{data}

\item{conf}{configurations}
}
\value{
Prameters to be provided to TMB
}
\description{
Set up data
}
\details{
define parameters
}
