% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{effetctiveFishingWidth}
\alias{effetctiveFishingWidth}
\title{effetctiveFishingWidth (not used in paper)}
\usage{
effetctiveFishingWidth(length)
}
\arguments{
\item{length}{in cm}
}
\value{


Returns the length dependent catchability constant
}
\description{
Returns the effective fishing width for a specific 5 cm length group.
}
