% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{fitModelSim}
\alias{fitModelSim}
\title{Simulation study, run models}
\usage{
fitModelSim(run, simData)
}
\arguments{
\item{run}{Run retruned by fitModel}

\item{simData}{One set of simulated data}
}
\value{
Returns a STIM object fitted with the simulated data
}
\description{
Re-estimate the model for each sampled data set returned by simStudy.
}
