% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{includeDummyAge}
\alias{includeDummyAge}
\title{Includes dummy variables for age}
\usage{
includeDummyAge(data)
}
\arguments{
\item{data}{Data for catch-at-length model internally used by TMB}
}
\description{
Includes dummy variables for age
}
\details{
This is just a dummy function to include empty variables in age-at-length model.
}
