% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{includeDummyPar}
\alias{includeDummyPar}
\title{includeDummyPar}
\usage{
includeDummyPar(par)
}
\arguments{
\item{par}{Parameters for the age-at-length model}
}
\value{
This function includes dummy parameters for age-at-length model when age data is not used.
}
\description{
Includes dummy-parameters if ALK is not applied.
}
