% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{jit}
\alias{jit}
\title{jit}
\usage{
jit(run, njit, ncores = 1, sd = 0.1)
}
\arguments{
\item{run}{The result of running fitModel}

\item{njit}{Number of jitter runs}

\item{ncores}{Number of cores to use}

\item{sd}{Standard deviation of noise to starting values}
}
\value{
A list with fitted runs and maximum difference in parameter estimates
}
\description{
This model takes the fitted returned by \code{\link{fitModel}}, and do a jitter-analysis to verify that the model is not sensitive to starting values.
}
