% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotResults}
\alias{plotResults}
\title{plotResults Plot results}
\usage{
plotResults(
  run,
  what = NULL,
  year = NULL,
  age = NULL,
  length = NULL,
  lon_lat = NULL,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL
)
}
\arguments{
\item{run}{fitted object returned by \code{\link{fitModel}}}

\item{what}{What to plot. Currently the four different options are implemented: 1) sunAlt: plots sun altitude effect. 2) depth: Plot depth effect. 3) ALK: plot ALK. 4) space: Spatial plot of catch-per-unit effort.}

\item{year}{Year of interest. Only relevant for ALK plots and spatial CPUE plots}

\item{age}{Age of interest for spatial CPUE plot}

\item{length}{Length of interest for spatial CPUE plot}

\item{lon_lat}{Longitude and latitude of interest for age-length-key plot}

\item{xlim}{optional xlim sent to fields::image.plot}

\item{ylim}{optional xlim sent to fields::image.plot}

\item{zlim}{optional zlim sent to fields::image.plot}
}
\description{
plotResults Plot results
}
