% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{pointsInStratum}
\alias{pointsInStratum}
\title{pointsInStratum}
\usage{
pointsInStratum(strataNumbers, stratas, locations)
}
\value{


A vector with index of points inside stratas providede in strataNumbers
}
\description{
pointsInStratum
}
