% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorMean.R
\name{posteriorMean}
\alias{posteriorMean}
\title{Posterior mean}
\usage{
posteriorMean(run)
}
\arguments{
\item{run}{The fitted model}
}
\description{
Posterior mean
}
\details{
Return posterior mean of the index (using the bias-correction flag in TMB::sdreport for the catch-at-length random effects only)
}
