% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{readRstoxFiles}
\alias{readRstoxFiles}
\title{readRstoxFiles}
\usage{
readRstoxFiles(conf)
}
\arguments{
\item{conf}{Configurations, only years to include used.}
}
\description{
creates a list containing the data files from stox.
}
