% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk_functions.R
\name{setMap_alk}
\alias{setMap_alk}
\title{setMap_alk}
\usage{
setMap_alk(conf_alk, par_alk)
}
\arguments{
\item{conf_alk}{Configurations}

\item{par_alk}{Parameters}
}
\value{
Returns the map-variable
}
\description{
Set up map-variable based on the configurations for the age-at-length model.
}
\details{
Sets up the map-variable
}
