% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{setPar}
\alias{setPar}
\title{setPar}
\usage{
setPar(data, conf)
}
\arguments{
\item{data}{Data}

\item{conf}{Configurations}
}
\value{
A list containing initial values for all model parameters and random effects in the model.
}
\description{
Initial values for model parameters and random effects for the catch-at-length model.
}
\details{
This function sets up the initial values for the parameters
}
