% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setupData}
\alias{setupData}
\title{setupData}
\usage{
setupData(dat_l, conf_l, confPred)
}
\arguments{
\item{dat_l}{A list of the data object returned by getBaseline, where each element of the list is a year}

\item{conf_l}{Configurations returned by \code{\link{defConf}}}

\item{confPred}{Prediction configurations returned by \code{\link{defConfPred}}}
}
\value{
This function return the length part of the data-list used by TMB
}
\description{
This function sets up the length part of the data-list used by TMB.
}
\details{
Set up data used by model
}
