% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk_functions.R
\name{setUpData_alk}
\alias{setUpData_alk}
\title{setUpData_alk}
\usage{
setUpData_alk(dat_alk, conf_alk, conf_l = NULL)
}
\arguments{
\item{dat_alk}{age data, a data table with one line for each fish}

\item{conf_alk}{configurations for age-at-length model}

\item{conf_l}{configurations for length-at-age model}
}
\value{
Data to be provided to TMB
}
\description{
Set up age data in the format needed by TMB.
}
\details{
Prepare data list for the age part. This data-list will be merged with the length data-list later.

The column dat_alk$readability provides information about what dat_alk$age refers to. If dat_alk$readability is missing, all ages are assumed to be measured correctly. If provided; 1: The corresponding age is assumed to be measured correctly. 5: The corresponding age is either dat_alk$age or dat_alk$age + 1.  6: the age is either dat_alk$age or dat_alk$age - 1. Any other number: Age information is not used
}
