% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simStudy}
\alias{simStudy}
\title{Simulation study}
\usage{
simStudy(run, nsim = 5)
}
\arguments{
\item{run}{Run retruned by fitModel}

\item{nsim}{Number of simulations}
}
\value{
A runSim object containtin all runs based on simulated data
}
\description{
Simulation study
}
