% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveIndices.R
\name{write_covariance_matrices}
\alias{write_covariance_matrices}
\title{Save indices in standard ICES format}
\usage{
write_covariance_matrices(run, file, digits = 6)
}
\arguments{
\item{run}{stim object}

\item{file}{Name of the file. Must end with .dat}

\item{digits}{Number of digits in the covariance matrix}
}
\description{
Save indices in standard ICES format
}
\details{
Saves a list with estimated yearly covariance matrices of the log-index.
}
