% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveIndices.R
\name{write_covriance_matrices}
\alias{write_covriance_matrices}
\title{Save indices in standard ICES format}
\usage{
write_covriance_matrices(run, file)
}
\arguments{
\item{run}{stim object}

\item{file}{Name of the file. Must end with .rds}
}
\description{
Save indices in standard ICES format
}
\details{
Saves a list with estimated yearly covariance matrices of the log-index.
}
