% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveIndices.R
\name{write_indices_ICES_format}
\alias{write_indices_ICES_format}
\title{Save indices in standard ICES format}
\usage{
write_indices_ICES_format(
  run,
  file,
  name,
  sdrep_bc = NULL,
  variance = FALSE,
  digits = 1
)
}
\arguments{
\item{run}{stim object}

\item{file}{Name of file. Must end with .dat}

\item{name}{name of index.}

\item{sdrep_bc}{sdreport from TMB with bias.correct}

\item{variance}{Save variances of log-indices?}

\item{digits}{number of decimal places in the output file}
}
\description{
Save indices in standard ICES format
}
\details{
Saves the estimated index or associated variance of log-index in standard ICES format
}
