% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defConf}
\alias{defConf}
\title{defConf}
\usage{
defConf(
  years,
  skipYears = NULL,
  spatial = 1,
  spatioTemporal = 0,
  nugget = 1,
  splineDepth = c(6, 1),
  sunAlt = c(1, 0),
  maxLength = NULL,
  dLength = 1,
  minLength = NULL,
  reduceLength = 3,
  cutoff = 100,
  cbound = 200,
  pcPriorRange = c(100, 0.1),
  pcPriorsd = c(1, 0.1),
  usePcPriors = 0,
  zeroInflated = 0,
  obsModel = 2,
  rwBeta0 = 1,
  applyALK = 0,
  mapRhoL = c(0, 1, 2),
  simulateProcedure = 1,
  stratasystem = list(),
  minDepth = 50,
  maxDepth = 600,
  trawlWidth = 1,
  plusGroup = 1
)
}
\arguments{
\item{years}{Years of data included}

\item{skipYears}{Years to skip in the estimation}

\item{spatial}{If 0-no spatial effect, 1-Include spatial effect}

\item{spatioTemporal}{If 0-no spatial-temporal effect, 1-Include spatial-temporal effect, 2-Include spatial-temporal effect with independence between years}

\item{nugget}{If 0-no nugget effect, 1-Include nugget effect}

\item{splineDepth}{Vector with depth configurations, first element is the k-variable in spline (6 means 6 basis function). Second element: 0- no depth effect, 1-depth effectk, 2-length dependent depth effect.}

\item{sunAlt}{Vector with sun altitude configurations, first element is the numberof basis functions. Second element: 0-sun effect, 1-length dependent sun effect.}

\item{maxLength}{Maximum length, used when defining the pluss group}

\item{minLength}{Minimum length, used when defining the minus group (probably never used)}

\item{reduceLength}{The resolution in length dimension used in latent effect}

\item{cutoff}{Cutoff used in mesh, first element is for spatial effect, second for spatio-temporal effect, third is 0 if same mesh is used for both spatial and spatio-temporal effect}

\item{cbound}{cbound used in mesh, first element is for spatial effect, second for spatio-temporal effect}

\item{pcPriorRange}{pc-priors for spatial and spatio-temporal effect}

\item{pcPriorsd}{pc-priors for spatial and spatio-temporal effect}

\item{usePcPriors}{If 0-No pc-priors used, 1- use pc-priors (probably never used)}

\item{zeroInflated}{If 0-No zero inflation, 1- use zero inflation (NB!! not fully implemented)}

\item{stratasystem}{List containing dsn and layer of shapefile}

\item{minDepth}{Minimum depth considered}

\item{maxDepth}{Maximum depth considered}

\item{trawlWidth}{Swept width of trawl}

\item{plusGroup}{Include plus goup? 1:yes, 0: No}
}
\description{
Configurations used. The object returned is
used in \code{\link{setupData}} and in  \code{\link{setPar}}.
}
