% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defConf}
\alias{defConf}
\title{defConf}
\usage{
defConf(
  years,
  lengthGroups,
  spatial = 1,
  spatioTemporal = 1,
  nugget = 1,
  beta0 = 1,
  splineDepth = c(6, 1),
  sunAlt = c(1, 0),
  duration = 0,
  maxLength = 100,
  minLength = 5,
  reduceLength = 3,
  cutoff = c(100, 100, 0),
  cbound = c(400, 100),
  pcPriorRange = c(100, 0.5),
  pcPriorsd = c(0.5, 0.5),
  usePcPriors = 0,
  zeroInflated = 0,
  obsModel = 2,
  mapRhoL = c(0, 1, 2),
  simulateProcedure = 1
)
}
\arguments{
\item{years}{Years of data included}

\item{lengthGroups}{Length groups included}

\item{spatial}{If 0-no spatial effect, 1-Include spatial effect}

\item{spatioTemporal}{If 0-no spatial-temporal effect, 1-Include spatial-temporal effect}

\item{nugget}{If 0-no nugget effect, 1-Include nugget effect}

\item{beta0}{0-no intercept (probably not used), 1- intercept included}

\item{splineDepth}{Vector with depth configurations, first element is the k-variable in spline (6 means 6 basis function). Second element: 0- no depth effect, 1-depth effectk, 2-length dependent depth effect.}

\item{sunAlt}{Vector with sun altitude configurations, first element is the numberof basis functions. Second element: 0-sun effectk, 1-length dependent sun effect.}

\item{duration}{Duration effect, 0- used as offset, 1- used as covariate.}

\item{maxLength}{Maximum length, used when defining the pluss group}

\item{minLength}{Minimum length, used when defining the minus group (probably never used)}

\item{reduceLength}{The resolution in length dimension used in latent effect}

\item{cutoff}{Cutoff used in mesh, first element is for spatial effect, second for spatio-temporal effect, third is 0 if same mesh is used for both spatial and spatio-temporal effect}

\item{cbound}{cbound used in mesh, first element is for spatial effect, second for spatio-temporal effect}

\item{pcPriorRange}{pc-priors for spatial and spatio-temporal effect}

\item{pcPriorsd}{pc-priors for spatial and spatio-temporal effect}

\item{usePcPriors}{If 0-No pc-priors used, 1- use pc-priors (probably never used)}

\item{zeroInflated}{If 0-No zero inflation, 1- use zero inflation (NB! not fully implemented)}
}
\description{
Configurations used. The object returned is
used in \code{\link{setupData}} and in  \code{\link{setPar}}.
}
