% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk_functions.R
\name{defConf_alk}
\alias{defConf_alk}
\title{defConf_alk}
\usage{
defConf_alk(
  years = NULL,
  minAge,
  maxAge,
  spatioTemporal = 0,
  spatial = 0,
  rwBeta0 = 0,
  betaLength = 0,
  meshSimilar = TRUE,
  readability = 1,
  usePCpriorsALK = 0,
  pcPriorsALKRange = c(300, 0.1),
  pcPriorsALKSD = c(1, 0.1)
)
}
\arguments{
\item{years}{Years to include}

\item{minAge}{minimum age}

\item{maxAge}{max age}

\item{spatioTemporal}{include spatio-tempora: 0-no 1-yes 2- yes but without correlation structure in time}

\item{spatial}{include spatial effect: 0-no 1-yes}

\item{rwBeta0}{random walk for intercelt within cohorts:  0-no 1-yes}

\item{betaLength}{Time varying regression coefficients for length: 0-no, 1-yes}

\item{meshSimilar}{If TRUE; apply the same mesh as in the package spatitemporalIndices, this is typically the case}

\item{readability}{If 1: Utilize age reading quality, if 0: Do not utilize age reading quality}

\item{usePCpriorsALK}{Use pc-priors: 0-no 1-yes. Probably never used; in case it is used, we need set pcPriorsALKRange and pcPriorsALKSD to reasonable values}

\item{pcPriorsALKRange}{pc-priors for range}

\item{pcPriorsALKSD}{pc-priors for standard deviations}
}
\value{
Configurations to set up the age-at-length part of the model
}
\description{
Set up the configurations for the age-at-length model
}
\details{
Defines the configurations
}
