% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defConfPred}
\alias{defConfPred}
\title{defConfPred}
\usage{
defConfPred(conf, nIntPoints = 4000, Depth = "Data")
}
\arguments{
\item{conf}{Configurations used when fitting the model}

\item{nIntPoints}{provide number of integration points}

\item{Depth}{if "NOAA": use NOAA data base for estimating depth in integration points}
}
\description{
Configurations used for prediction.
}
