% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{defConfPred}
\alias{defConfPred}
\title{defConfPred}
\usage{
defConfPred(strata = 1:26, nIntPoints = 4000)
}
\arguments{
\item{strata}{Which statas to predict indices in (probably always all stratas and extract more detailed predictions on a later stage)}

\item{yearToPredict}{Prediciton year}

\item{conf}{Configurations used when fitting the model}

\item{n}{provide number of integration points and number of bootstrap samples in internal validation routine}
}
\description{
Configurations used for prediction.
}
