% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk_functions.R
\name{defpar_alk}
\alias{defpar_alk}
\title{defpar_alk}
\usage{
defpar_alk(data_alk, conf_alk)
}
\arguments{
\item{data_alk}{data-list for the age-at-length model}

\item{conf_alk}{configurations for the age-at-length model}
}
\value{
The age-at-length model parameter list with initial values.
}
\description{
Set up the age-at-length model parameter list used by TMB.
}
\details{
Set up the age-at-length model parameter list used by TMB.
}
