% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{fitModel}
\alias{fitModel}
\title{Run model}
\usage{
fitModel(
  dat_l,
  conf_l,
  confPred,
  dat_alk = NULL,
  conf_alk = NULL,
  parSet = NULL,
  runModel = TRUE,
  mapSet = NULL,
  intern = FALSE,
  twoStage = FALSE,
  ...
)
}
\arguments{
\item{dat_l}{Data frame with length data}

\item{confPred}{Configurations for predictions}

\item{dat_alk}{Data frame with age data}

\item{conf_alk}{Configurations for ALK model.}

\item{...}{Parameters sendt to sdreport.}

\item{conf}{Configurations for length part of model}
}
\value{
A fitted stim object. The indices with corresponding uncertainties can be extracted by using the \code{\link{saveIndex}} function. Or manually by inspecting the fit$rl object.
}
\description{
This funtion runs the index model given the data on catch-at-length and age-at-length.
Indices with corresponding covariance matrices and uncertianty estimates can be extracted using the \code{\link{saveIndex}} function.
}
\details{
This model runs estimate the index model, and returns the fitted model.
}
