% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{fitModel}
\alias{fitModel}
\title{Run model}
\usage{
fitModel(
  dat_l,
  conf_l,
  confPred,
  dat_alk = NULL,
  conf_alk = NULL,
  parPrior = NULL,
  low = list(),
  up = list()
)
}
\arguments{
\item{dat_l}{Data frame with length data}

\item{confPred}{Configurations for predictions}

\item{dat_alk}{Data frame with age data}

\item{conf_alk}{Configurations for ALK model}

\item{low}{Lower bounds in the optimization, typically not used.}

\item{up}{Upper bounds in the optimization, typically not used.}

\item{conf}{Configurations for length part of model}
}
\value{
A fitted stim object
}
\description{
Run model
}
