% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{fitModel}
\alias{fitModel}
\title{Run model}
\usage{
fitModel(
  dat_l,
  conf_l,
  confPred,
  dat_alk = NULL,
  conf_alk = NULL,
  parSet = NULL,
  mapSet = NULL,
  runModel = TRUE,
  twoStage = FALSE,
  ...
)
}
\arguments{
\item{dat_l}{Data frame with length data}

\item{conf_l}{Configurations for length part of model}

\item{confPred}{Configurations for predictions}

\item{dat_alk}{Data frame with age data}

\item{conf_alk}{Configurations for ALK model.}

\item{parSet}{Initial values; only used if not using default values (only for experimental use)}

\item{mapSet}{Map-list; Used to overwrite default Map-list. (only for experimental use)}

\item{runModel}{Estimate the model if TRUE, if not we only set up the model}

\item{twoStage}{Estimate the model in two steps. May reduce computation time if we use temporal correlation in spatio-temporal effect}

\item{...}{Parameters sent to TMB::sdreport.}
}
\value{
A fitted stim object. The indices with corresponding uncertainties can be extracted by using the \code{\link{saveIndex}} function. Or manually by inspecting the fit$rl object.
}
\description{
This funtion runs the index model given the data on catch-at-length and age-at-length.
Indices with corresponding covariance matrices and uncertianty estimates can be extracted using the \code{\link{saveIndex}} function.
}
\details{
This model runs estimate the index model, and returns the fitted model.
}
