% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{fitModel}
\alias{fitModel}
\title{Run model}
\usage{
fitModel(
  data,
  par,
  conf,
  map = NULL,
  doDetailedRep = 2,
  low = list(),
  up = list()
)
}
\arguments{
\item{data}{Data}

\item{par}{Parameters}

\item{conf}{Configurations}

\item{doDetailedRep}{Return reported indices etc if true.}

\item{low}{Bounds}

\item{up}{Bounds}
}
\value{
A fitted stim object
}
\description{
Run model
}
