% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{includeDummyPar}
\alias{includeDummyPar}
\title{Includes dummy-parameters if ALK is not applied.}
\usage{
includeDummyPar(par)
}
\arguments{
\item{par}{Parameters}
}
\value{
This function includes dummy parameters when ALK in not applied.
}
\description{
Includes dummy-parameters if ALK is not applied.
}
