% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{retroSTIM}
\alias{retroSTIM}
\title{retroSTIM}
\usage{
retroSTIM(run, nyears, ncores = 1)
}
\arguments{
\item{run}{The result of running fitModel}

\item{nyears}{The number of years to remove sequentially}

\item{ncores}{Number of cores to use.}
}
\value{
The fitted model within each peal in the retrospective analysis
}
\description{
This function do a retrospective analysis and returns the fitted models within each peal
}
