% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveIndices.R
\name{saveIndex}
\alias{saveIndex}
\title{Save indices with assosiated variance and covariance}
\usage{
saveIndex(run, file, folder = "", sdrep_bc = NULL)
}
\arguments{
\item{run}{stim object}

\item{file}{name of file with index.}

\item{folder}{Path to the folder the index is saved.}

\item{sdrep_bc}{Output of TMB::sdreport(..., bias.correct = TRUE)}
}
\description{
Save indices with assosiated variance and covariance
}
\details{
Saves the esitmated index with assosiated variance estimates and covariance structures in a format that can be directly utilized by SAM.
}
