% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{setMap}
\alias{setMap}
\title{setMap}
\usage{
setMap(par, conf)
}
\arguments{
\item{par}{Parameters included}

\item{conf}{Configurations}
}
\description{
\code{setMap} defines the map-argument used in MakeADFun.
}
