% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk_functions.R
\name{setMap_alk}
\alias{setMap_alk}
\title{Set up map variable}
\usage{
setMap_alk(conf, par)
}
\arguments{
\item{conf}{Configurations}

\item{par}{Parameters}
}
\value{
Returns the map-variable
}
\description{
Set up map variable
}
\details{
Sets up the map-variable
}
