% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setupData}
\alias{setupData}
\title{setupData}
\usage{
setupData(dataLength, conf, confPred)
}
\arguments{
\item{conf}{Configurations, see \code{\link{defConf}} for details.}

\item{confPred}{Prediction configurations}

\item{data}{A list of the data object returned by getBaseline, where each element of the list is a year}
}
\description{
setupData
}
\details{
Set up data used by model
}
