% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setupData}
\alias{setupData}
\title{setupData}
\usage{
setupData(data = NULL, conf, confPred = NULL, useProj = FALSE, ...)
}
\arguments{
\item{data}{A list of the data object returned by getBaseline, where each element of the list is a year}

\item{conf}{Configurations, see \code{\link{defConf}} for details.}

\item{confPred}{Prediction configurations}
}
\value{

}
\description{
setupData
}
\details{
Set up data used by model
}
